Identified issues in the check_if_closed function. Basically it was giving false values due to the forward/reversed settings in the tecplot streamline generation function. This quirk actually should resolve any error in streamlines going to the 'wrong' pole.
The created surface still appears extremely jagged and I believe it's
due to either the 'ordered zone' logic in tecplot, the way the
streamzone data is being fed in, or both.

Moving the R_TAIL_CAP up to -20Re from -30 improves the total coverage
of the created streamzones, something to consider as things come along.