updated interpolation for surface construction function to actually interpolate when data is missing directly in the angle bin between the curve set of points.
got plasmasheet to plot well for poorly formed test case, had to adjust
far end of boundary to x_min +5 in order to get enough points since the
boundary is comprised of just a couple stream lines. Maybe adjusting the
critera to >1x tail cap for the closed criteria could help this.