cleaned up plasmasheet function to work with latest tools functions.
Modified the number of spoke points in surface construction to 10*alpha
rather than 2*nalpha. Seemed to improve interpolated curve. Should make
this into keyword argument with default.

Plasmasheet 3D surface is not centered around Z/Y=0 axis so initializing
the zone to a cylinder with this central axis creates artifacts. Need to
modify center point at each slice to be the z/y  average. Possibly store
this in stream_points.csv as new item to read in.