was overwritting interface variables each time instead of assigning each to it's particular zone.
giving up on trying to get the lobes as a single iso-surface. it may be
possible with updated Tecplot versions which allow zonemerging, would
have to manipulate XYZ coordinates to trick tecplot into thinking
north and south lobes are connected

^ Instead going back to north and south lobes and merge during
extraction. It is suboptimal but will keep analysis from growing